refactor: remove 'edit user post' component and add edit button to 'user post' component
