--- angular/src/app/components/user-post-components/edit-user-post/edit-user-post.component.css
@@ -1,3 +1,0 @@
-button {
-  margin: 10px;
-}
