--- angular/src/app/components/user-post-components/edit-user-post/edit-user-post.component.html
@@ -1,1 +1,0 @@
-<button *ngIf="userPost" mat-flat-button (click)="openDialog()">Edit</button>
