+++ angular/src/app/components/user-post-components/user-post-container/user-post-container.component.css
@@ -6,3 +6,7 @@
 h2 {
   color: white;
 }
+
+button {
+  margin: 10px;
+}
