+++ angular/src/app/components/user-post-components/user-post-container/user-post-container.component.html
@@ -2,6 +2,9 @@
   <h2>User Post Container</h2>
   <app-user-post-form (onAddUserPost)="addUserPost($event)"></app-user-post-form>
   <div class="user-posts" *ngFor="let userPost of userPosts">
-    <app-user-post [userPost]="userPost" (onDeleteUserPost)="deleteUserPost($event)"></app-user-post>
+    <app-user-post
+      [userPost]="userPost"
+      (onDeleteUserPost)="deleteUserPost($event)"
+      (onUpdateUserPost)="updateUserPost($event)"></app-user-post>
   </div>
 </div>
