+++ angular/src/app/components/user-post-components/user-post/user-post.component.html
@@ -3,5 +3,5 @@
   <h3>{{userPost.id}}: {{userPost.user.firstName}} {{userPost.user.lastName}}</h3>
   <p>{{userPost.description}}</p>
   <button mat-flat-button type="button" (click)="deleteUserPost()">Delete</button>
-  <app-edit-user-post [userPost]="userPost"></app-edit-user-post>
+  <button *ngIf="userPost" mat-flat-button (click)="openDialog()">Edit</button>
 </div>
